+++ src/lua/config_argmap.lua
@@ -39,13 +39,13 @@ package.cpath = os.getenv("LUA_CPATH") or
 -- LuaLogging: A simple API to use logging features in Lua: https://neopallium.github.io/lualogging/manual.html#introduction
 
 -- For html links, use relative paths (to test/output folder) - more portable
-DIR_MJS_JSON = os.getenv("DIR_MJS_JSON") or "mapjs-json"
+DIR_MAPJS_JSON = os.getenv("DIR_MAPJS_JSON") or "mapjs-json"
 
 -- QUESTION: Should I be using a path join function?
-PATH_FILE_CONFIG_MJS = os.getenv("PATH_FILE_CONFIG_MJS")
-PATH_FILE_CONFIG_MJS_PROCESSED = os.getenv("PATH_FILE_CONFIG_MJS_PROCESSED")
+PATH_FILE_CONFIG_MAPJS = os.getenv("PATH_FILE_CONFIG_MAPJS")
+PATH_FILE_CONFIG_MAPJS_PROCESSED = os.getenv("PATH_FILE_CONFIG_MAPJS_PROCESSED")
 DIR_PUBLIC_OUTPUT = os.getenv("DIR_PUBLIC_OUTPUT") or (config.project_folder .. "/test/output")
-PATH_DIR_PUBLIC_MJS_JSON = os.getenv("PATH_DIR_PUBLIC_MJS_JSON") or (DIR_PUBLIC_OUTPUT .. "/" .. DIR_MJS_JSON)
+PATH_DIR_PUBLIC_MAPJS_JSON = os.getenv("PATH_DIR_PUBLIC_MAPJS_JSON") or (DIR_PUBLIC_OUTPUT .. "/" .. DIR_MAPJS_JSON)
 DIR_HTML_SERVER_OUTPUT = os.getenv("DIR_HTML_SERVER_OUTPUT") or "output"
 
 PATH_DIR_LAYOUTS = os.getenv("PATH_DIR_LAYOUTS") or config.project_folder .. "/src/layouts" -- Reads the container and controls html