+++ docs/CHANGELOG.md
@@ -6,6 +6,13 @@
   - Add note about linking/using templates (html and latex) with pandoc.
   - Add references to argmap specs spreadsheet?
 
+## argmap 15.0.0 BREAKING
+
+- BREAKING: Rename all config settings from MJS to MAPJS  for clarity and consistency e.g. `PATH_FILE_CONFIG_MAPJS` and `PATH_FILE_CONFIG_MAPJS_PROCESSED` from `PATH_FILE_CONFIG_MJS` and `PATH_FILE_CONFIG_MJS_PROCESSED`.
+- `mapjs/src`: Remove `yaml-loader!` directive from require yaml requests, now webpack has correct loader configuration.
+
+BREAKING: All such variable names needs to be updated in local `mapjs/config/environment-mapjs.yaml` and `config/environment-argmap.yaml`, consistent with the new defaults file.
+
 ## argmap 14.0.2
 
 - `webpack.config.js`: